chore: standards-conformance remediation pass (2026-07-05)
Restores docs/standards/ACCESSIBILITY-STANDARD.md byte-identical to the
pinned v1.0.1 upstream — it had been edited in-repo with self-flattering
claims (e.g. asserting 2.4.12 AAA conformance and a shipped LHCI/contrast
gate) instead of being upstreamed as a proposal. The honest proposal is
now tracked separately in docs/standards-proposal-2026-07-05-accessibility.md.

Also lands, per audit-2026-07-05/gtfs-scorecard-REMEDIATION.md:
- P0: branch-protection ruleset + tag ruleset artifacts (not yet applied
  live), SECURITY.md SLA rewrite, README standards-conformance/
  observability/versioning declarations, version-of-truth unification
  (pyproject/CITATION.cff/server.json all 1.0.0) with a CI drift check.
- P1: dependency-audit + CodeQL (python/actions) + zizmor + TruffleHog +
  OpenSSF Scorecard workflows, CHANGELOG.md, wheel build in CI, ci.yml
  now runs a single `make verify`.
- Quick wins: pre-commit hook version bumps, ruff select widened (98
  findings resolved, 16 pre-existing complexity findings ratcheted via
  docs/lint-complexity-ratchet.md rather than silenced), CODEOWNERS,
  duplicate ADR number fixed (0013 -> 0030), otp-qa image pinned by
  digest, PR template / CONTRIBUTING pointed at `make verify`.

Housekeeping alongside this pass: the orphaned `v1`/`v1.0.0` tags (found
to point at a commit not an ancestor of main's history) were local-only
and have been deleted; they were never pushed, so no remote cleanup is
needed.

`make verify` stayed green throughout. See the Execution Log in
audit-2026-07-05/gtfs-scorecard-REMEDIATION.md for full detail, including
a caught-and-reverted accidental `pre-commit run --all-files` that briefly
touched ~3,280 generated files (confirmed not present in this commit).

+++ pipeline/src/scorecard_pipeline/badge.py
@@ -48,7 +48,7 @@ _PAD = 8.0
 
 
 def _segment_width(text: str) -> int:
-    return int(round(len(text) * _CHAR_PX + 2 * _PAD))
+    return round(len(text) * _CHAR_PX + 2 * _PAD)
 
 
 def render_badge(grade: str, score: float | None = None, expiry_status: str | None = None) -> str:
